removed old vagrant
